ui.frontend/src/utility/hooks/__tests__/useCreateFetcher.test.ts
import { useCreateFetcher } from "../useCreateFetcher";
import { renderHook } from "@testing-library/react";
jest.mock('src/providers/LoggingProvider');
//import { useLogger } from 'src/providers/LoggingProvider';

// Mocking hooks used inside useCreateFetcher


// Test cases for useCreateFetcher function

// Describe block encapsulates all the test cases related to useCreateFetcher function
describe("useCreateFetcher", () => {

  beforeEach(()=>{
    // Setting up mock logger and output message

  });

  afterEach(()=>{

  })

  // Test Case: Check whether useCreateFetcher returns object with 3 properties
  test("useCreateFetcher should return an object containing 3 properties", () => {
    const { result } = renderHook(() => useCreateFetcher());
    expect(Object.keys(result.current).length).toBe(3);
    expect('createGraphQLFetcher' in result.current).toBe(true);
    expect('createMultiLanguageFetcher' in result.current).toBe(true);
    expect('createAPIFetcher' in result.current).toBe(true);
  });

  // Test Case: Check whether 'createGraphQLFetcher' works without throwing error
  test("useCreateFetcher should not throw error for createGraphQLFetcher", () => {
    const { result } = renderHook(() => useCreateFetcher());
    const { createGraphQLFetcher } = result.current;
    expect(() => createGraphQLFetcher({ url: "" })).not.toThrow();
  });

  // Test Case: Check whether 'createMultiLanguageFetcher' works without throwing error
  test("useCreateFetcher should not throw error for createMultiLanguageFetcher", () => {
    const { result } = renderHook(() => useCreateFetcher());
    const { createMultiLanguageFetcher } = result.current;
    expect(() => createMultiLanguageFetcher({ url: "" })).not.toThrow();
  });

  // Test Case: Check whether 'createAPIFetcher' works without throwing error
  test("useCreateFetcher should not throw error for createAPIFetcher", () => {
    const { result } = renderHook(() => useCreateFetcher());
    const { createAPIFetcher } = result.current;
    expect(() => createAPIFetcher({ url: "" })).not.toThrow();
  });

  // ... other individual test cases to cover all aspects of the functions

});